backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
google-generativeai>=0.8.3
temporalio>=1.9.0
google-genai
agno>=2.5.0
sqlalchemy
llama-index-core>=0.14.0
llama-index-embeddings-google-genai>=0.1.0
langfuse>=2.0.0
neo4j>=5.0.0
